backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
langchain-openai==0.0.5
langchain-google-genai==0.0.6
pydantic==2.5.0
sse-starlette==1.6.5
openai==1.3.5
python-dotenv==1.0.0
starlette==0.27.0